refactor(formatter): move format cases out of lib
